Create function playRound - compares user input with getComputerChoice
Proceed to use an if..else statement for this. It's quite long but I tested it via the console and it works. I'm planning to add another function that will serve as the start 'button' for the round. Or more accurately, a start function since the GUI will be worked later on.