Create game() function which counts wins and outputs winner
Managed to utilize for loop to count wins. I did many revisions and finally got the gist of the code after multiple failed attempts. A lot of errors surfaced during the process but I was able to iron them out by utilizing what I've learned so far. I could've used boolean to count the wins by returning a false value for every lost but I instead used slice(), which determined if the first part of the win text is 'Win' or 'lose' and went from there.